#22 move swagger files from "/" to "/swagger"
