+++ src/main/resources/static/swagger/css/print.css
