+++ src/main/resources/static/swagger/css/reset.css
