+++ src/main/resources/static/swagger/css/screen.css
