+++ src/main/resources/static/swagger/css/style.css
