+++ src/main/resources/static/swagger/css/typography.css
