+++ src/main/resources/static/swagger/fonts/droid-sans-v6-latin-regular.ttf
