+++ src/main/resources/static/swagger/images/explorer_icons.png
