+++ src/main/resources/static/swagger/images/favicon-16x16.png
