+++ src/main/resources/static/swagger/images/favicon-32x32.png
