+++ src/main/resources/static/swagger/images/favicon.ico
