+++ src/main/resources/static/swagger/images/logo_small.png
