+++ src/main/resources/static/swagger/images/pet_store_api.png
