+++ src/main/resources/static/swagger/images/throbber.gif
