+++ src/main/resources/static/swagger/images/wordnik_api.png
