+++ src/main/resources/static/swagger/index.html
@@ -27,7 +27,7 @@
       if (url && url.length > 1) {
         url = decodeURIComponent(url[1]);
       } else {
-        url = "http://localhost:8080/api-docs";
+        url = "/api-docs";
       }
       window.swaggerUi = new SwaggerUi({
         url: url,
