+++ src/main/resources/static/swagger/lib/handlebars-2.0.0.js
