+++ src/main/resources/static/swagger/lib/swagger-oauth.js
