+++ src/main/resources/static/swagger/o2c.html
