+++ src/main/resources/static/swagger/swagger-ui.js
